Refactor operate tests and usage of __tests__ folder for logic
