+++ src/logic/__tests__/operate.test.js
@@ -3,35 +3,39 @@ import operate from '../operate';
 it('makes basic math operations', () => {
   const a = Math.floor(Math.random() * 10);
   const b = Math.floor(Math.random() * 10);
-  const plus = '+';
-  const minus = '-';
-  const times = 'x';
-  const by = '÷';
-  const modulo = '%';
-  const sum = operate(a, b, plus);
-  expect(sum).toBe(`${a + b}`);
-  const sub = operate(a, b, minus);
-  expect(sub).toBe(`${a - b}`);
-  const mult = operate(a, b, times);
-  expect(mult).toBe(`${a * b}`);
-  const division = parseFloat(operate(a, b, by));
-  expect(division).toBe((a / b));
-  const mod = operate(a, b, modulo);
-  expect(mod).toBe(`${a % b}`);
-});
+  const operators = ['+', '-', 'x', '÷'];
+  function notInfinite(operator) {
+    expect(operate(a, b, operator)).not.toBe(Infinity);
+  }
+  operators.forEach((operator) => {
+    const result = operate(a, b, operator);
+    switch (operator) {
+      case '+':
+        expect(result).toBe(`${a + b}`);
+        notInfinite(operator);
+        break;
+      case '-':
+        expect(result).toBe(`${a - b}`);
+        notInfinite(operator);
+        break;
+      case 'x':
+        expect(result).toBe(`${a * b}`);
+        notInfinite(operator);
+        break;
+      case '÷':
+        if (b === 0) {
+          expect(result).toBe("Can't divide by 0.");
+        } else {
+          expect(`${parseFloat(result)}`).toBe(`${a / b}`);
+          notInfinite(operator);
+        }
 
-it('will make only the right operations', () => {
-  const a = Math.floor(Math.random() * 10);
-  const b = 0;
-  const plus = '+';
-  const minus = '-';
-  const times = 'x';
-  const by = '÷';
-  const modulo = '%';
-  const sum = operate(a, b, plus);
-  console.log(sum);
-  expect(sum).not.toMatch(`${a + b * plusOrMinus()} `);
-})
+        break;
+      default:
+        expect(result).toBe(0);
+    }
+  });
+});
 
 it('will crash if there is a division by 0', () => {
   const a = Math.floor(Math.random() * 10);
@@ -42,9 +46,6 @@ it('will crash if there is a division by 0', () => {
     expect(crash).toBe("Can't divide by 0.");
   } else {
     const division = parseFloat(operate(a, b, by));
-    expect(division).toBe(parseFloat(a / b));
-  }
-  if (b === 0) {
-    expect(crash).toBe('Can\'t divide by 0.');
+    expect(division).toBe((a / b));
   }
 });
